Update software template to support janus-idp argocd plugin
The PR replaces the previous label of the component registered in
backstage with a label-selector that is supported by janus-idp argocd
plugin.

In addition, it labels the source repository for showing the CD tab with
the argocd application that used for the CI (tekton pipelines are
triggered by argocd and both can be presented in the context of the
source repository).

Signed-off-by: Moti Asayag <masayag@redhat.com>

rh-pre-commit.version: 2.3.1
rh-pre-commit.check-secrets: ENABLED